MuxBaseExoPlayer class split to two versions 2_96 and later and prior to 2_9_6
